requirements.txt
flask
cloudinary
pillow
rembg
gunicorn
